Move ProxyComponentResource into Python SDK
